token init in a function
